# demo.openapi.universita/src/main/java/demo/openapi/universita/controller/generated/StudentiApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.19.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package demo.openapi.universita.controller.generated;

import java.util.List;
import java.util.Optional;

import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.context.request.NativeWebRequest;

import demo.openapi.universita.model.ErroreImprevisto;
import demo.openapi.universita.model.Studente;
import demo.openapi.universita.model.StudenteCreate;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import jakarta.validation.constraints.NotNull;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-02-01T22:35:36.136170+01:00[Europe/Rome]", comments = "Generator version: 7.19.0")
@Validated
@Tag(name = "studenti", description = "the studenti API")
public interface StudentiApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_AGGIORNA_STUDENTE = "/studenti/{id}";
    /**
     * PUT /studenti/{id} : Aggiorna Studente
     * Aggiorna i campi di uno Studente in base al suo ID
     *
     * @param id Identificativo (ID) di un unico Studente (required)
     * @param studente  (required)
     * @return Risposta con esito positivo e lo Studente aggiornato (status code 200)
     *         or Richiesta incompleta o errata (status code 400)
     *         or Risorsa non trovata (status code 404)
     *         or Errore generico (status code 200)
     */
    @Operation(
        operationId = "aggiornaStudente",
        summary = "Aggiorna Studente",
        description = "Aggiorna i campi di uno Studente in base al suo ID",
        responses = {
            @ApiResponse(responseCode = "200", description = "Risposta con esito positivo e lo Studente aggiornato", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Studente.class))
            }),
            @ApiResponse(responseCode = "400", description = "Richiesta incompleta o errata", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroreImprevisto.class))
            }),
            @ApiResponse(responseCode = "404", description = "Risorsa non trovata", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroreImprevisto.class))
            }),
            @ApiResponse(responseCode = "default", description = "Errore generico", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroreImprevisto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = StudentiApi.PATH_AGGIORNA_STUDENTE,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Studente> aggiornaStudente(
        @NotNull @Parameter(name = "id", description = "Identificativo (ID) di un unico Studente", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "Studente", description = "", required = true) @Valid @RequestBody Studente studente
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"cognome\" : \"cognome\", \"corsoDiLaurea\" : \"corsoDiLaurea\", \"sesso\" : \"N\", \"nome\" : \"nome\", \"id\" : 0, \"codiceFiscale\" : \"codiceFiscale\", \"matricola\" : \"matricola\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"messaggio\" : \"messaggio\", \"dettaglio\" : \"dettaglio\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"messaggio\" : \"messaggio\", \"dettaglio\" : \"dettaglio\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"messaggio\" : \"messaggio\", \"dettaglio\" : \"dettaglio\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_CANCELLA_STUDENTE = "/studenti/{id}";
    /**
     * DELETE /studenti/{id} : Cancella Studente
     * Cancella uno Studente in base al suo ID
     *
     * @param id Identificativo (ID) di un unico Studente (required)
     * @return Nessun contenuto (status code 204)
     *         or Richiesta incompleta o errata (status code 400)
     *         or Risorsa non trovata (status code 404)
     *         or Errore generico (status code 200)
     */
    @Operation(
        operationId = "cancellaStudente",
        summary = "Cancella Studente",
        description = "Cancella uno Studente in base al suo ID",
        responses = {
            @ApiResponse(responseCode = "204", description = "Nessun contenuto"),
            @ApiResponse(responseCode = "400", description = "Richiesta incompleta o errata", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroreImprevisto.class))
            }),
            @ApiResponse(responseCode = "404", description = "Risorsa non trovata", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroreImprevisto.class))
            }),
            @ApiResponse(responseCode = "default", description = "Errore generico", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroreImprevisto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = StudentiApi.PATH_CANCELLA_STUDENTE,
        produces = { "application/json" }
    )
    default ResponseEntity<Void> cancellaStudente(
        @NotNull @Parameter(name = "id", description = "Identificativo (ID) di un unico Studente", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"messaggio\" : \"messaggio\", \"dettaglio\" : \"dettaglio\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"messaggio\" : \"messaggio\", \"dettaglio\" : \"dettaglio\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"messaggio\" : \"messaggio\", \"dettaglio\" : \"dettaglio\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_CERCA_STUDENTE = "/studenti/{id}";
    /**
     * GET /studenti/{id} : Cerca Studente
     * Restituisce uno Studente in base al suo ID
     *
     * @param id Identificativo (ID) di un unico Studente (required)
     * @return Risposta con esito positivo e lo Studente trovato (status code 200)
     *         or Richiesta incompleta o errata (status code 400)
     *         or Risorsa non trovata (status code 404)
     *         or Errore generico (status code 200)
     */
    @Operation(
        operationId = "cercaStudente",
        summary = "Cerca Studente",
        description = "Restituisce uno Studente in base al suo ID",
        responses = {
            @ApiResponse(responseCode = "200", description = "Risposta con esito positivo e lo Studente trovato", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Studente.class))
            }),
            @ApiResponse(responseCode = "400", description = "Richiesta incompleta o errata", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroreImprevisto.class))
            }),
            @ApiResponse(responseCode = "404", description = "Risorsa non trovata", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroreImprevisto.class))
            }),
            @ApiResponse(responseCode = "default", description = "Errore generico", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroreImprevisto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = StudentiApi.PATH_CERCA_STUDENTE,
        produces = { "application/json" }
    )
    default ResponseEntity<Studente> cercaStudente(
        @NotNull @Parameter(name = "id", description = "Identificativo (ID) di un unico Studente", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"cognome\" : \"cognome\", \"corsoDiLaurea\" : \"corsoDiLaurea\", \"sesso\" : \"N\", \"nome\" : \"nome\", \"id\" : 0, \"codiceFiscale\" : \"codiceFiscale\", \"matricola\" : \"matricola\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"messaggio\" : \"messaggio\", \"dettaglio\" : \"dettaglio\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"messaggio\" : \"messaggio\", \"dettaglio\" : \"dettaglio\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"messaggio\" : \"messaggio\", \"dettaglio\" : \"dettaglio\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_CREA_STUDENTE = "/studenti";
    /**
     * POST /studenti : Crea Studente
     * Creazione di uno Studente.
     *
     * @param studenteCreate  (required)
     * @return Risorsa creata: Studente (status code 201)
     *         or Richiesta incompleta o errata (status code 400)
     *         or Errore generico (status code 200)
     */
    @Operation(
        operationId = "creaStudente",
        summary = "Crea Studente",
        description = "Creazione di uno Studente.",
        responses = {
            @ApiResponse(responseCode = "201", description = "Risorsa creata: Studente", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Studente.class))
            }),
            @ApiResponse(responseCode = "400", description = "Richiesta incompleta o errata", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroreImprevisto.class))
            }),
            @ApiResponse(responseCode = "default", description = "Errore generico", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroreImprevisto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = StudentiApi.PATH_CREA_STUDENTE,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Studente> creaStudente(
        @Parameter(name = "StudenteCreate", description = "", required = true) @Valid @RequestBody StudenteCreate studenteCreate
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"cognome\" : \"cognome\", \"corsoDiLaurea\" : \"corsoDiLaurea\", \"sesso\" : \"N\", \"nome\" : \"nome\", \"id\" : 0, \"codiceFiscale\" : \"codiceFiscale\", \"matricola\" : \"matricola\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"messaggio\" : \"messaggio\", \"dettaglio\" : \"dettaglio\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"messaggio\" : \"messaggio\", \"dettaglio\" : \"dettaglio\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_LISTA_STUDENTI = "/studenti";
    /**
     * GET /studenti : Lista Studenti
     * Restituisce tutti gli Studenti
     *
     * @return Risposta con esito positivo e uno o piu Studenti trovati (status code 200)
     *         or Errore generico (status code 200)
     */
    @Operation(
        operationId = "listaStudenti",
        summary = "Lista Studenti",
        description = "Restituisce tutti gli Studenti",
        responses = {
            @ApiResponse(responseCode = "200", description = "Risposta con esito positivo e uno o piu Studenti trovati", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Studente.class)))
            }),
            @ApiResponse(responseCode = "default", description = "Errore generico", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErroreImprevisto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = StudentiApi.PATH_LISTA_STUDENTI,
        produces = { "application/json" }
    )
    default ResponseEntity<List<Studente>> listaStudenti(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"cognome\" : \"cognome\", \"corsoDiLaurea\" : \"corsoDiLaurea\", \"sesso\" : \"N\", \"nome\" : \"nome\", \"id\" : 0, \"codiceFiscale\" : \"codiceFiscale\", \"matricola\" : \"matricola\" }, { \"cognome\" : \"cognome\", \"corsoDiLaurea\" : \"corsoDiLaurea\", \"sesso\" : \"N\", \"nome\" : \"nome\", \"id\" : 0, \"codiceFiscale\" : \"codiceFiscale\", \"matricola\" : \"matricola\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"messaggio\" : \"messaggio\", \"dettaglio\" : \"dettaglio\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}